Implemented besser exception output.
git-svn-id: svn://svn.code.sf.net/p/captainsengine/code/trunk@653 dc34a4ac-3e42-4537-918b-2dea8b01751e

+++ Engine/io/MemReader.h
@@ -29,7 +29,7 @@ public:
   virtual bool isWorking();
   virtual void skip(int numberOfElements);
 
-  virtual uint32 getSize() {return mEnd-mStart;}
+  virtual uint32 getSize() {return (uint32)(mEnd-mStart);}
 
   char* getData() {return mStart;}
   char* getCurrentData() {return mCurrent;}
